+++ src/Components/ConnectForm/ConnectForm.module.scss
@@ -1,0 +1,33 @@
+.connect {
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    height: 91px;
+
+    &__control {
+        display: flex;
+        width: 515px;
+        @media screen and (max-width: 750px) {
+            width: 450px;
+        }
+    }
+    input {
+        width: 60%;
+        height: 47px;
+        border-radius: 0 0 0 10px;
+        outline: none;
+        font-weight: 300;
+        font-size: 24px;
+        border: 1px solid #2b3a67;
+        box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+    }
+    button {
+        width: 40%;
+        height: 47px;
+    }
+}
+
+.highlight {
+    font-weight: 700;
+    color: #66999b;
+}
