+++ src/pages/MainPage/MainPage.module.scss
@@ -68,54 +68,11 @@
         height: 47px;
     }
 }
-.connect {
-    display: flex;
-    justify-content: space-between;
-    flex-direction: column;
-    height: 91px;
 
-    &__control {
-        display: flex;
-        width: 515px;
-        @media screen and (max-width: 750px) {
-            width: 450px;
-        }
-    }
-    input {
-        width: 60%;
-        height: 47px;
-        border-radius: 0 0 0 10px;
-        outline: none;
-        font-weight: 300;
-        font-size: 24px;
-        border: 1px solid #2b3a67;
-        box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
-    }
-    button {
-        width: 40%;
-        height: 47px;
-    }
-}
 .center {
     text-align: center;
 }
-.highlight {
-    font-weight: 700;
-    color: #66999b;
-}
 
-.blue {
-    background-color: blue;
-}
 
-.error {
-    width: 300px;
-    height: 50px;
-    font-weight: 700;
-    font-size: 26px;
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    letter-spacing: 1px;
-    color: #9f5800;
-}
+
+
